sw_redlinehide_3: fix SetRedlineFlags() skipping over redlines
If there's a insert redline inside a footnote whose anchor is in a
delete redline, the nested redline will be copied by the first
iteration of the loop in SetRedlineFlags() and then deleted in the
second iteration, while handling the outer delete redline.

This causes the redline immediately following the outer delete
redline to be skipped in the 2nd iteration.

Change-Id: Iba5c3a7b0d6d8a0d76a6ab6c7066cc4b059edb75